refactor: fallback component for reuse
